refactor(analyzer): split project.rs into project, pass2, and diagnostics
project.rs (747 lines) keeps ProjectAnalyzer orchestration, AnalysisResult,
and file-discovery helpers. pass2.rs (808 lines) introduces Pass2Driver<'a>
which owns all body-analysis methods (analyze_bodies, analyze_fn_decl,
analyze_class_decl, and their *_typed variants, plus merge_return_types).
diagnostics.rs (211 lines) holds the free diagnostic helpers
(check_type_hint_classes, check_name_class, emit_unused_params,
emit_unused_variables, offset_to_line_col).

merge_return_types and collect_php_files are re-exported from project.rs so
existing call sites in expr.rs and composer.rs are unchanged.

Note: the parallel per-file loop stays inside ProjectAnalyzer::analyze()
rather than moving to Pass2Driver::analyze_pass_2(); pulling it in would
bundle cache lookup and the on_file_done callback into pass2, muddying the
boundary without meaningful gain.
