+++ crates/mir-analyzer/src/lib.rs
@@ -4,10 +4,12 @@ pub mod class;
 pub mod collector;
 pub mod context;
 pub mod dead_code;
+pub mod diagnostics;
 pub mod expr;
 pub mod generic;
 pub mod narrowing;
 pub mod parser;
+pub mod pass2;
 pub mod php_version;
 pub mod project;
 pub mod stmt;
